+++ santashop-app/src/app/features/pre-registration/profile/profile.form.ts
@@ -1,10 +1,4 @@
-import {
-	AbstractControl,
-	FormControl,
-	FormGroup,
-	ValidationErrors,
-	Validators,
-} from '@angular/forms';
+import { FormControl, FormGroup, Validators } from '@angular/forms';
 
 export interface ChangeEmailForm {
 	password: FormControl<string | undefined>;
@@ -14,7 +8,6 @@ export interface ChangeEmailForm {
 export interface ChangePasswordForm {
 	oldPassword: FormControl<string | undefined>;
 	newPassword: FormControl<string | undefined>;
-	newPassword2: FormControl<string | undefined>;
 }
 
 const validators = {
@@ -59,30 +52,13 @@ export const changeEmailForm = (): FormGroup<ChangeEmailForm> =>
 	});
 
 export const changePasswordForm = (): FormGroup<ChangePasswordForm> =>
-	new FormGroup<ChangePasswordForm>(
-		{
-			oldPassword: new FormControl(undefined, {
-				nonNullable: true,
-				validators: validators.password,
-			}),
-			newPassword: new FormControl(undefined, {
-				nonNullable: true,
-				validators: validators.password,
-			}),
-			newPassword2: new FormControl(undefined, {
-				nonNullable: true,
-				validators: validators.password,
-			}),
-		},
-		{ validators: passwordMatchValidator },
-	);
-
-function passwordMatchValidator(
-	control: AbstractControl,
-): ValidationErrors | null {
-	const password = control.get('newPassword')?.value;
-	const confirmation = control.get('newPassword2')?.value;
-	return password && confirmation && password !== confirmation
-		? { passwordMismatch: true }
-		: null;
-}
+	new FormGroup<ChangePasswordForm>({
+		oldPassword: new FormControl(undefined, {
+			nonNullable: true,
+			validators: validators.password,
+		}),
+		newPassword: new FormControl(undefined, {
+			nonNullable: true,
+			validators: validators.password,
+		}),
+	});